refactor(dashboard): Integrate sidebar, filter and search functionality
